InstantAdjusters: Moves `plus` and `minus` to adjuster class
And removed `InstantTruncator` completely
